requirements.txt
pytesseract==0.3.10
Pillow==10.3.0
pandas==2.2.2
streamlit==1.31.0